Add task/experiment sweep dispatchers
- split forecasting sweep train/eval logic into task adapters under src/workflows/sweeps/tasks/forecasting
- turn src/workflows/sweeps/{train,eval}.py into task-family/experiment-type dispatchers
- extend canonical sweep shell launchers with TASK_FAMILY and EXPERIMENT_TYPE
- keep existing chronos2 sweep wrappers compatible while routing through dispatcher
- update architecture docs, scripts guide, and project tracking for the taxonomy pivot
