+++ project_tracking.csv
@@ -40,5 +40,5 @@ wandb-mirror,P3,pending,Optional W&B mirror for paper artifacts,,"Optional W&B m
 repo-design-handbook-ratification,P1,done,Ratify repository design handbook,scripts-reorg-and-prune,"Completed 2026-08-14: Phase F governance workstream ratified `docs/architecture/repository-design-handbook.md` decisions (`src/workflows/` top-level package, Python-first orchestrators with thin shell wrappers, and canonical pre-MLflow run-manifest schema). Delivered PR-F1/PR-F2/PR-F3/PR-F2b/PR-F3b/PR-F4, including canonical run-manifest v1 writer (`src/workflows/runtime/manifest.py`) and manifest emission in representative training (`src/workflows/forecasting/pipeline.py`) and evaluation (`scripts/experiments/nocturnal_hypo_eval.py`) entrypoints.",,,
 phase-f-ratification-checklist-plan,P1,done,Create Phase F ratification checklist PR plan,,"Completed 2026-08-14: created `agent_notes/P1-37_PHASE_F_RATIFICATION_CHECKLIST_PR_PLAN.md` with PR-F1..PR-F4 sequence, validation gates, and exit criteria for ratification rollout.",,,
 workflows-package-migration,P1,in_progress,Create src/workflows package and migrate pipelines,repo-design-handbook-ratification,"Started 2026-08-14: top-level `src/workflows/` package created and workflow runtime moved to canonical `src/workflows/forecasting/pipeline.py`; temporary holdout compatibility wrappers were removed after call-site migration. Remaining: migrate additional multi-step pipelines under ratified taxonomy.",,,
-python-first-orchestration-transition,P1,in_progress,Transition sweep orchestration to Python-first,repo-design-handbook-ratification,"Started 2026-08-14: PR-F3 baseline completed with Python-first Chronos-2 sweep training orchestrator plus thin shell wrapper and preserved env-var contract. PR-F3b added a model-agnostic sweep training core (`src/workflows/sweeps/train.py`) and generic CLI (`scripts/experiments/sweep_train.py`) with Chronos-2 training profile (`configs/experiments/nocturnal_forecast/chronos2_forecasting_train_sweep.yaml`). Updated 2026-08-16: added model-agnostic sweep evaluation core (`src/workflows/sweeps/eval.py`) and generic eval CLI (`scripts/experiments/sweep_eval.py`) plus canonical taxonomy-aligned shell launchers (`scripts/training/sweeps/run_sweep_train.sh`, `scripts/evaluation/sweeps/run_sweep_eval.sh`). Removed model-specific Chronos-2 Python profile wrappers and retained only thin compatibility shell wrappers in `scripts/experiments/`. Remaining: migrate additional model-family sweep/eval shell surfaces to the same config-driven generic orchestration pattern and continue script-surface relocation out of `scripts/experiments`.",,,
+python-first-orchestration-transition,P1,in_progress,Transition sweep orchestration to Python-first,repo-design-handbook-ratification,"Started 2026-08-14: PR-F3 baseline completed with Python-first Chronos-2 sweep training orchestrator plus thin shell wrapper and preserved env-var contract. PR-F3b added generic sweep CLIs (`scripts/experiments/sweep_train.py`, `scripts/experiments/sweep_eval.py`) and canonical shell launchers (`scripts/training/sweeps/run_sweep_train.sh`, `scripts/evaluation/sweeps/run_sweep_eval.sh`). Updated 2026-08-16 (taxonomy pivot slice): sweep CLIs now dispatch by `task-family` + `experiment-type`, and forecasting/nocturnal_forecast logic was moved into task adapters (`src/workflows/sweeps/tasks/forecasting/{train,eval}.py`). This establishes a multi-task extension point for future families (e.g., event detection) while preserving current forecasting behavior. Remaining: add additional task/experiment adapters and continue script-surface relocation out of `scripts/experiments`.",,,
 holdout-boundary-cleanup,P1,done,Complete forecasting workflow boundary cleanup,workflows-package-migration,"Completed 2026-08-14: canonical namespace moved to `src/workflows/forecasting/`; active execution path now uses extracted modeling/evaluation modules (`src/workflows/forecasting/modeling.py`, `src/workflows/forecasting/evaluation.py`); legacy duplicate implementations removed from the orchestrator pipeline; holdout compatibility modules/import surfaces retired.",,,
