+++ scripts/experiments/chronos2_sweep_eval.sh
@@ -14,6 +14,8 @@ PROJECT_ROOT="$(cd "$SCRIPT_DIR/../.." && pwd)"
 cd "$PROJECT_ROOT"
 
 MODEL_TYPE="chronos2" \
+TASK_FAMILY="${TASK_FAMILY:-forecasting}" \
+EXPERIMENT_TYPE="${EXPERIMENT_TYPE:-nocturnal_forecast}" \
 SWEEP_SPEC="configs/experiments/nocturnal_forecast/chronos2_forecasting_eval_sweep.yaml" \
 EVAL_PYTHON="${EVAL_PYTHON:-.venvs/autogluon/bin/python}" \
 bash scripts/evaluation/sweeps/run_sweep_eval.sh "$@"
