+++ src/workflows/sweeps/tasks/__init__.py
@@ -1,0 +1,1 @@
+"""Task-family sweep adapters."""
